// frontend/src/api/generated/model/transportConfigApplyRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { TransportConfigApplyRequestOperation } from './transportConfigApplyRequestOperation';
import type { TransportLinkedOutboundEnsure } from './transportLinkedOutboundEnsure';
import type { TransportSpec } from './transportSpec';

export interface TransportConfigApplyRequest {
  operation: TransportConfigApplyRequestOperation;
  transport: TransportSpec;
  linked_outbound: TransportLinkedOutboundEnsure;
}